(PC-35028)[PRO] refactor: extract address select handler in dedicated util
	+ enable custom address handler in AddressSelect component
