+++ pro/src/components/Address/Address.tsx
@@ -5,6 +5,7 @@ import { useDebouncedCallback } from 'use-debounce'
 import { getDataFromAddress } from 'apiClient/adresse/apiAdresse'
 import { AdresseData, FeaturePropertyType } from 'apiClient/adresse/types'
 import { SelectOption } from 'commons/custom_types/form'
+import { handleAddressSelect } from 'commons/utils/handleAddressSelect'
 import { normalizeStrForAdressSearch } from 'commons/utils/searchPatternInOptions'
 import { serializeAdressData } from 'components/Address/serializer'
 import { SelectAutocomplete } from 'ui-kit/form/SelectAutoComplete/SelectAutocomplete'
@@ -17,6 +18,11 @@ interface AddressProps {
   onlyTypes?: FeaturePropertyType[]
   disabled?: boolean
   className?: string
+  customHandleAddressSelect?: (
+    setFieldValue: any,
+    selectedItem?: AutocompleteItemProps,
+    searchField?: FieldInputProps<string>
+  ) => void
 }
 
 export interface AutocompleteItemProps {
@@ -32,6 +38,7 @@ export const AddressSelect = ({
   onlyTypes,
   disabled = false,
   className,
+  customHandleAddressSelect,
 }: AddressProps) => {
   const { setFieldValue } = useFormikContext()
   const [options, setOptions] = useState<SelectOption[]>([])
@@ -43,6 +50,7 @@ export const AddressSelect = ({
   useEffect(() => {
     setOptions([{ label: selectedField.value, value: selectedField.value }])
   }, [selectedField.value])
+  const addressSelectHandler = customHandleAddressSelect ?? handleAddressSelect
 
   const onSearchFieldChange = async () => {
     if (searchField.value.length >= 3) {
@@ -64,7 +72,7 @@ export const AddressSelect = ({
       )
     } else if (searchField.value.length === 0 && searchFieldMeta.touched) {
       setOptions([])
-      handleAddressSelect(setFieldValue, undefined, searchField)
+      addressSelectHandler(setFieldValue, undefined, searchField)
     }
   }
 
@@ -78,7 +86,7 @@ export const AddressSelect = ({
     // False positive, eslint disable can be removed when noUncheckedIndexedAccess is enabled in TS config
     // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
     if (addressesMap[searchField.value] !== undefined) {
-      handleAddressSelect(
+      addressSelectHandler(
         setFieldValue,
         addressesMap[searchField.value],
         searchField
@@ -126,20 +134,4 @@ export const AddressSelect = ({
       className={className}
     />
   )
-}
-
-export const handleAddressSelect = (
-  setFieldValue: any,
-  selectedItem?: AutocompleteItemProps,
-  searchField?: FieldInputProps<string>
-) => {
-  setFieldValue('street', selectedItem?.extraData?.address ?? '')
-  if (searchField) {
-    setFieldValue('addressAutocomplete', searchField.value)
-  }
-  setFieldValue('postalCode', selectedItem?.extraData?.postalCode ?? '')
-  setFieldValue('city', selectedItem?.extraData?.city ?? '')
-  setFieldValue('latitude', selectedItem?.extraData?.latitude ?? '')
-  setFieldValue('longitude', selectedItem?.extraData?.longitude ?? '')
-  setFieldValue('banId', selectedItem?.value ?? '')
 }
