+++ pro/src/pages/VenueSettings/SiretOrCommentFields/SiretOrCommentFields.tsx
@@ -4,7 +4,7 @@ import { useState } from 'react'
 import { getDataFromAddress } from 'apiClient/adresse/apiAdresse'
 import { getSiretData } from 'commons/core/Venue/getSiretData'
 import { humanizeSiret, unhumanizeSiret } from 'commons/core/Venue/utils'
-import { handleAddressSelect } from 'components/Address/Address'
+import { handleAddressSelect } from 'commons/utils/handleAddressSelect'
 import { serializeAdressData } from 'components/Address/serializer'
 import { FormLayout } from 'components/FormLayout/FormLayout'
 import { TextArea } from 'ui-kit/form/TextArea/TextArea'
